Reorganizing package
